requirements.txt
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
eventlet==0.33.3
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
python-engineio==4.7.1
python-socketio==5.9.0
simple-websocket==0.10.1
six==1.16.0
urllib3==2.0.5
wsproto==1.2.0
yarl==1.9.2
